% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorrorPalettes.R
\name{horror_colors}
\alias{horror_colors}
\title{horror_cols}
\usage{
horror_colors(...)
}
\arguments{
\item{...}{Character names of horror_cols. Options include vintageblood, brightblood, blood, and darkblood.}
}
\description{
Function to extract various shades of blood as hex codes
}
\examples{
horror_colors('vintageblood')
horror_colors('brightblood', 'darkblood')
}
\keyword{colors,}
\keyword{theme}
