% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorrorPalettes.R
\docType{data}
\name{horror_cols}
\alias{horror_cols}
\title{horror_cols}
\format{An object of class \code{character} of length 4.}
\usage{
horror_cols
}
\description{
Assorted shades of blood. Any updates to this palette should be made within this object. This object is used by other functions; if you need to call hex codes directly use horror_colors.
}
\examples{
 horror_cols #shows all colors and codes
 horror_cols['blood']
}
\keyword{colors,}
\keyword{theme}
