% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorrorPalettes.R
\name{horror_palettes}
\alias{horror_palettes}
\title{horror_palettes}
\usage{
horror_palettes(palette = "GetOut", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in SH_pals. Default palette is Qualitative1. Options include sequential color palettes
(PrimaryGreen based on paradiso, SecondaryOrange based on Flamingo, TertiaryYellow, TertiaryGreen, TertiaryStormy, and Greyscale), diverging palettes (Divergent_PF, using paradiso and flamingo; Divergent_YS, using yellow and stormy), and qualitative palettes (Qualitative1 and Qualitative2)}

\item{reverse}{Boolean indicating whether the palette should be reversed. Default is FALSE}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a Spring Health color palette
}
\examples{
horror_palettes('GetOut')
}
\keyword{palettes,}
\keyword{scales,}
\keyword{theme}
