% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorrorPalettes.R
\docType{data}
\name{horror_pals}
\alias{horror_pals}
\title{horror_pals}
\format{An object of class \code{list} of length 17.}
\usage{
horror_pals
}
\description{
A list of horror movie inspired palettes.  Many palettes from colorpalette.cinema on Instagram. Any updates to these palettes should be made within this object. This object is used by other functions; if you need to call palettes directly use horror_palettes().
}
\examples{
 horror_pals['SOTL']
}
\keyword{palettes,}
\keyword{theme}
