% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorrorPalettes.R
\name{scale_color_horror}
\alias{scale_color_horror}
\title{scale_color_horror}
\usage{
scale_color_horror(palette = "Exorcist_Q", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in horror_pals. Default palette is Exorcist_Q.
Options include sequential color palettes (PrimaryGreen based on paradiso, SecondaryOrange based on Flamingo,
TertiaryYellow, TertiaryGreen, TertiaryStormy, and Greyscale), diverging palettes (Divergent_PF, using paradiso and flamingo;
Divergent_YS, using yellow and stormy), and qualitative palettes (Qualitative1 and Qualitative2)}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not. Default is TRUE.}

\item{reverse}{Boolean indicating whether the palette should be reversed. Default is FALSE.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
ggplot2 compatible function for selecting the color scale based on Spring Health palettes
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
 geom_point(size = 4) +
 scale_color_horror(reverse = F, 'Nosferatu_Q')

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  geom_point(size = 4, alpha = .6) +
  scale_color_horror(discrete = F, palette = 'NeonDemon_Divergent')
}
\keyword{palettes,}
\keyword{scales,}
\keyword{theme}
