% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorrorPalettes.R
\name{scale_fill_horror}
\alias{scale_fill_horror}
\title{scale_fill_horror}
\usage{
scale_fill_horror(palette = "Exorcist_Q", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in horror_palettes. Options: "Exorcist", "GetOut", "NeonDemon_blue",
"NeonDemon_red", "TheShining", "SOTL", "Suspiria", "Excision_Divergent", "GetOut_Divergent", "NeonDemon_Divergent",
DawnoftheDead_Q", "Excision_Q", "Exorcist_Q", "Nosferatu_Q", "Suspiria_Q", "TexasChainsaw_Q", "Greyscale".}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
ggplot2 compatible function for selecting the fill scale based on Spring Health palettes
}
\examples{
library(ggplot2)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
geom_bar() +
theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
scale_fill_horror(palette = "TheShining", guide = 'none')
}
