% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorrorPalettes.R
\name{show_me_the_palette}
\alias{show_me_the_palette}
\title{show_me_the_palette}
\usage{
show_me_the_palette(palette)
}
\arguments{
\item{palette}{Character name of palette in horror_palettes. Options: "Exorcist", "GetOut", "NeonDemon_blue",
"NeonDemon_red", "TheShining", "SOTL", "Suspiria", "Excision_Divergent", "GetOut_Divergent", "NeonDemon_Divergent",
DawnoftheDead_Q", "Excision_Q", "Exorcist_Q", "Nosferatu_Q", "Suspiria_Q", "TexasChainsaw_Q", "Greyscale".}
}
\description{
A function that will print the colors for any one of the available palettes.
This function only works on a single palette at a time
}
\examples{
show_me_the_palette("GetOut")
show_me_the_palette("Nosferatu_Q")
}
